pyproject.toml
[project]
name = "recsystrx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asttokens==3.0.0",
    "colorama==0.4.6",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "exceptiongroup==1.2.2",
    "executing==2.2.0",
    "faiss-cpu>=1.11.0",
    "filelock==3.18.0",
    "fsspec==2025.3.2",
    "implicit>=0.7.2",
    "ipykernel==6.29.5",
    "ipython==8.36.0",
    "ipywidgets>=8.1.6",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.5",
    "packaging==25.0",
    "pandas>=2.2.3",
    "parso==0.8.4",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "polars==1.29.0",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pygments==2.19.1",
    "pynndescent>=0.5.13",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pywin32==310",
    "pyzmq==26.4.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "six==1.17.0",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "torch>=2.7.0",
    "torchtext>=0.18.0",
    "torchvision>=0.22.0",
    "tornado==6.4.2",
    "tqdm>=4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.13.2",
    "wcwidth==0.2.13",
]
